+++ xclim/testing/tests/test_indices.py
@@ -283,6 +283,22 @@ class TestAgroclimaticIndices:
         if method == "icclim":
             np.testing.assert_array_equal(bedd, bedd_high_lat)
 
+    def test_cool_night_index(self):
+        ds = open_dataset("cmip5/tas_Amon_CanESM2_rcp85_r1i1p1_200701-200712.nc")
+        ds = ds.rename(dict(tas="tasmin"))
+
+        cni = xci.cool_night_index(tasmin=ds.tasmin, lat=ds.lat)
+        tasmin = convert_units_to(ds.tasmin, "degC")
+
+        cni_nh = cni.where(cni.lat >= 0, drop=True)
+        cni_sh = cni.where(cni.lat < 0, drop=True)
+
+        tn_nh = tasmin.where((tasmin.lat >= 0) & (tasmin.time.dt.month == 9), drop=True)
+        tn_sh = tasmin.where((tasmin.lat < 0) & (tasmin.time.dt.month == 3), drop=True)
+
+        np.testing.assert_array_equal(cni_nh, tn_nh)
+        np.testing.assert_array_equal(cni_sh, tn_sh)
+
 
 class TestDailyFreezeThawCycles:
     def test_simple(self, tasmin_series, tasmax_series):
@@ -1957,3 +1973,72 @@ def test_wind_chill(tas_series, sfcWind_series):
 
     out = xci.wind_chill_index(tas=tas, sfcWind=sfcWind, method="US")
     assert out[-1].isnull()
+
+
+class TestPotentialEvapotranspiration:
+    def test_baier_robertson(self, tasmin_series, tasmax_series):
+        tn = tasmin_series(np.array([0, 5, 10]) + 273.15)
+        tn = tn.expand_dims(lat=[45])
+        tx = tasmax_series(np.array([10, 15, 20]) + 273.15)
+        tx = tx.expand_dims(lat=[45])
+
+        out = xci.potential_evapotranspiration(tn, tx, method="BR65")
+        np.testing.assert_allclose(out[0, 2], [3.861079 / 86400], rtol=1e-2)
+
+    def test_hargreaves(self, tasmin_series, tasmax_series, tas_series):
+        tn = tasmin_series(np.array([0, 5, 10]) + 273.15)
+        tn = tn.expand_dims(lat=[45])
+        tx = tasmax_series(np.array([10, 15, 20]) + 273.15)
+        tx = tx.expand_dims(lat=[45])
+        tm = tas_series(np.array([5, 10, 15]) + 273.15)
+        tm = tm.expand_dims(lat=[45])
+
+        out = xci.potential_evapotranspiration(tn, tx, tm, method="HG85")
+        np.testing.assert_allclose(out[2, 0], [3.962589 / 86400], rtol=1e-2)
+
+    def test_thornthwaite(self, tas_series):
+        time_std = date_range(
+            "1990-01-01", "1990-12-01", freq="MS", calendar="standard"
+        )
+        tm = xr.DataArray(
+            np.ones((time_std.size, 1)),
+            dims=("time", "lat"),
+            coords={"time": time_std, "lat": [45]},
+            attrs={"units": "degC"},
+        )
+
+        out = xci.potential_evapotranspiration(tas=tm, method="TW48")
+        np.testing.assert_allclose(out[0, 1], [42.7619242 / (86400 * 30)], rtol=1e-1)
+
+
+def test_water_budget(pr_series, tasmin_series, tasmax_series):
+    pr = pr_series(np.array([10, 10, 10]))
+    pr.attrs["units"] = "mm/day"
+    pr = pr.expand_dims(lat=[45])
+    tn = tasmin_series(np.array([0, 5, 10]) + K2C)
+    tn = tn.expand_dims(lat=[45])
+    tx = tasmax_series(np.array([10, 15, 20]) + K2C)
+    tx = tx.expand_dims(lat=[45])
+
+    out = xci.water_budget(pr, tn, tx, method="BR65")
+    np.testing.assert_allclose(out[0, 2], [6.138921 / 86400], rtol=1e-3)
+
+    out = xci.water_budget(pr, tn, tx, method="HG85")
+    np.testing.assert_allclose(out[0, 2], [6.037411 / 86400], rtol=1e-3)
+
+    time_std = date_range("1990-01-01", "1990-12-01", freq="MS", calendar="standard")
+    tm = xr.DataArray(
+        np.ones((time_std.size, 1)),
+        dims=("time", "lat"),
+        coords={"time": time_std, "lat": [45]},
+        attrs={"units": "degC"},
+    )
+    prm = xr.DataArray(
+        np.ones((time_std.size, 1)) * 10,
+        dims=("time", "lat"),
+        coords={"time": time_std, "lat": [45]},
+        attrs={"units": "mm/day"},
+    )
+
+    out = xci.water_budget(prm, tas=tm, method="TW48")
+    np.testing.assert_allclose(out[1, 0], [8.5746025 / 86400], rtol=1e-1)